+++ log/%DATE%/.9539c437085f44fb8f20acf802aa4fce58ca8252-audit.json
@@ -39,6 +39,16 @@
             "date": 1728797538216,
             "name": "log/2024-10-13/http-2024-10-13.log.4",
             "hash": "ecc594bfcb451201d95950e2d2474b93bb66236191c4046eb48595913192faa9"
+        },
+        {
+            "date": 1729077207801,
+            "name": "log/2024-10-16/http-2024-10-16.log",
+            "hash": "a8ff04c7dbe7754efef0e5ff536d02941f775abaa0523b3e7030805326b330a6"
+        },
+        {
+            "date": 1729171390142,
+            "name": "log/2024-10-17/http-2024-10-17.log",
+            "hash": "b1694264c9188ca61a34a7698fddbe351162c732659ac69e8e256af18b1d12ba"
         }
     ],
     "hashType": "sha256"
