+++ src/common/auth.constant.ts
@@ -1,0 +1,3 @@
+export const AUTH_CONSTANT = {
+  HASH_SALT_ROUNDS: 10,
+};
