--- src/common/enums/enum-social-providers.ts
@@ -1,6 +1,0 @@
-export enum SocialProviders {
-  LOCAL = 'LOCAL',
-  GOOGLE = 'GOOGLE',
-  NAVER = 'NAVER',
-  KAKAO = 'KAKAO',
-}
