--- src/constants/auth.constant.ts
@@ -1,3 +1,0 @@
-export const AUTH_CONSTANT = {
-  HASH_SALT_ROUNDS: 10,
-};
