+++ src/enums/enum-social-providers.ts
@@ -1,0 +1,6 @@
+export enum SocialProviders {
+  LOCAL = 'LOCAL',
+  GOOGLE = 'GOOGLE',
+  NAVER = 'NAVER',
+  KAKAO = 'KAKAO',
+}
